file: split a "file" protocol into as a package
