+++ core/errors.go
@@ -1,6 +1,6 @@
 package core
 
 type httpError interface {
-	statusCode() int
-	body() string
+	StatusCode() int
+	Body() string
 }
