+++ internal/httperror/httperror.go
@@ -1,3 +1,5 @@
+// Package httperror provides error type which composed from HTTP status code
+// and message.
 package httperror
 
 import "fmt"
@@ -20,10 +22,10 @@ func (err httpErr) Error() string {
 	return fmt.Sprintf(err.format, err.args...)
 }
 
-func (err httpErr) statusCode() int {
+func (err httpErr) StatusCode() int {
 	return err.status
 }
 
-func (err httpErr) body() string {
+func (err httpErr) Body() string {
 	return err.Error()
 }
