+++ plugins/plugins.go
@@ -12,6 +12,7 @@ import (
 	_ "github.com/koron/nvgd/protocol/aws"
 	_ "github.com/koron/nvgd/protocol/configp"
 	_ "github.com/koron/nvgd/protocol/db"
+	_ "github.com/koron/nvgd/protocol/file"
 	_ "github.com/koron/nvgd/protocol/help"
 	_ "github.com/koron/nvgd/protocol/redis"
 	_ "github.com/koron/nvgd/protocol/version"
