--- protocol/util.go
@@ -1,5 +1,0 @@
-package protocol
-
-import "regexp"
-
-var rxLastComponent = regexp.MustCompile(`[^/]+/?$`)
